Refactor "VariantAnalysisHeader" component
So that we can re-use parts of it in the model alerts view
